Breaking up the schedule into controllers and other classes
